moved camera logic to createCat form
